+++ apps/client/app/(afterLogin)/(mobile)/mobile/attendance-check-input/_components/StudyInfo.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import { Text } from "@wow-class/ui";
+import useAttendanceCheckSearchParams from "hooks/useAttendanceCheckSearchParams";
+const StudyInfo = () => {
+  const { studyName, currentWeek } = useAttendanceCheckSearchParams();
+  return (
+    <>
+      <Text typo="body1">
+        {studyName}
+        {currentWeek}주차
+      </Text>
+    </>
+  );
+};
+
+export default StudyInfo;
