+++ apps/client/hooks/useAttendanceCheckSearchParams.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useSearchParams } from "next/navigation";
 
 const useAttendanceCheckSearchParams = () => {
